+++ pystella/rf/light_curve_func.py
@@ -175,7 +175,7 @@ def curves_read_mix(fname, dtype=None,
     band.Band.load_settings()
 
     if dtype is None:
-        dtype = [('time', np.float), ('filter', 'S8'), ('mag', np.float), ('err', np.float)]
+        dtype = [('time', float), ('filter', 'S8'), ('mag', float), ('err', float)]
 
     lc_data = np.loadtxt(fname, skiprows=skiprows, dtype=dtype, comments=comments)  # jd filter mag mage
     if is_out:
